Added test for Neutrino interface with files

+++ framework/CodeInterfaces/Neutrino/neutrinoInterface.py
@@ -10,11 +10,11 @@ import warnings
 warnings.simplefilter('default',DeprecationWarning)
 
 import os
-import xml.etree.ElementTree as ET
+import lxml.etree as ET
 
 from CodeInterfaceBaseClass import CodeInterfaceBase
 
-class NeutrinoBase(CodeInterfaceBase):
+class Neutrino(CodeInterfaceBase):
   """
     Provides code to interface RAVEN to Neutrino code
     The name of this class represents the type in the RAVEN input file
@@ -25,7 +25,7 @@ class NeutrinoBase(CodeInterfaceBase):
       </Code>
       ...
     </Models>
-	
+
   """
 
   def generateCommand(self, inputFiles, executable, clargs=None,fargs=None,preExec=None):
@@ -33,7 +33,7 @@ class NeutrinoBase(CodeInterfaceBase):
       See base class.  Collects all the clargs and the executable to produce the command-line call.
       Returns tuple of commands and base file name for run.
       Commands are a list of tuples, indicating parallel/serial and the execution command to use.
-      @ In, inputFiles, list, List of input files (length of the list depends on the number of inputs have 
+      @ In, inputFiles, list, List of input files (length of the list depends on the number of inputs have
 	        been added in the Step is running this code)
       @ In, executable, string, executable name with absolute path (e.g. /home/path_to_executable/code.exe)
       @ In, clargs, dict, optional, dictionary containing the command-line flags the user can specify in the input
@@ -44,25 +44,25 @@ class NeutrinoBase(CodeInterfaceBase):
      	     run the code (string), returnCommand[1] is the name of the output root
     """
     found = False
-	
+
     # Find the first file in the inputFiles that is an XML, which is what we need to work with.
     for index, inputFile in enumerate(inputFiles):
       if self._isValidInput(inputFile):
         found = True
         break
-    if not found: 
+    if not found:
       raise Exception('No correct input file has been found. Got: '+' '.join(inputFiles))
-	
+
     #Determines the path to the input file
     path = inputFiles[0].getAbsFile()
-    
+
     #Creates the output file that saves information that is outputted to the command prompt
 	#The output file name is results + Neutrino input file name.
 	#Example: resultsNeutrinoInput
     outputfile = 'results'
-    
+
     #Creates run command tuple (['executionType','execution command'], output file root)
-	#The path to the Neutrino executable is specified in the RAVEN input file as the executable 
+	#The path to the Neutrino executable is specified in the RAVEN input file as the executable
 	# since it must change directories to run
     returnCommand = [('serial','cd ' + executable + ' && Neutrino.exe --nogui --file ' + str(path) \
     + ' --run')], outputfile
@@ -76,9 +76,9 @@ class NeutrinoBase(CodeInterfaceBase):
       @ Out, valid, bool, 'True' if an input file has an extension of '.nescene', otherwise 'False'.
     """
     valid = False
-    if inputFile.getExt() in ('nescene'): 
+    if inputFile.getExt() in ('nescene'):
       valid = True
-    
+
     return valid
 
   def getInputExtension(self):
@@ -97,8 +97,8 @@ class NeutrinoBase(CodeInterfaceBase):
       @ In , currentInputFiles, list,  list of current input files (input files of this iteration)
       @ In , oriInputFiles, list, list of the original input files
       @ In , samplerType, string, Sampler type (e.g. MonteCarlo, Adaptive, etc. see manual Samplers section)
-      @ In , Kwargs, dictionary, kwarded dictionary of parameters. In this dictionary there is another 
-	         dictionary called "SampledVars" where RAVEN stores the variables that got sampled 
+      @ In , Kwargs, dictionary, kwarded dictionary of parameters. In this dictionary there is another
+	         dictionary called "SampledVars" where RAVEN stores the variables that got sampled
 			 (e.g. Kwargs['SampledVars'] => {'var1':10,'var2':40})
       @ Out, newInputFiles, list, list of newer input files, list of the new input files (modified and not)
     """
@@ -108,11 +108,11 @@ class NeutrinoBase(CodeInterfaceBase):
       if self._isValidInput(inputFile):
         found = True
         break
-    if not found: 
+    if not found:
       raise Exception('No correct input file has been found. Got: '+' '.join(oriInputFiles))
 
     originalPath = currentInputFiles[index].getAbsFile()
-    
+
     # Since the input file is XML we can load and edit it directly using etree
     # Load the XML into a tree:
     tree = ET.parse(originalPath, ET.XMLParser(encoding='utf-8'))
@@ -121,43 +121,43 @@ class NeutrinoBase(CodeInterfaceBase):
 
     # grep the variables that got sampled
     varDict = Kwargs['SampledVars']
-	
+
     # Go through sampled variables
     for var in varDict:
       #Search for the SPH solver properties
       #NIISphSolver_1 name may need to be changed based on Neutrino input file
-      #Can add other properties to change beside the solver properties 
+      #Can add other properties to change beside the solver properties
       for element in root.findall('./properties/Scene/NIISphSolver_1/'):
-        #Search for the Radius property 
+        #Search for the Radius property
         if element.get('name') == 'ParticleSize':
           #Set the radius value to the sampled value
           element.set('val',str(varDict[var]))
 
 
-        
+
         #Change where the measurements and the output data is stored in the input file to match RAVEN location
         #Search for the Base properties
         for elementBase in root.findall('./properties/Base/'):
           #Search for the SceneFilePath property
           if elementBase.get('name') == 'SceneFilePath':
-            #Set the SceneFilePath 
+            #Set the SceneFilePath
             elementBase.set('val', str(originalPath))
-            
+
           #Search for the SaveDir property
           if elementBase.get('name') == 'SaveDir':
             #Create and set SaveDir
 			#NeutrinoInput.nescene needs to be changed to the Neutrino input file name
             savePath = originalPath.replace("NeutrinoInput.nescene","",1)
             elementBase.set('val',str(savePath))
-                
+
           if elementBase.get('name') == 'CacheDir':
             #Create and set CacheDir
 			#NeutrinoInput.nescene needs to be changed to the Neutrino input file name
             cachePath = originalPath.replace("NeutrinoInput.nescene","",1)
             elementBase.set('val',str(cachePath))
-        
+
         #Search for the Measurement field properties
-		#MeasurementField_1 name may need to be changed based on Neutrino input file 
+		#MeasurementField_1 name may need to be changed based on Neutrino input file
         for elementMeas in root.findall('./properties/Scene/MeasurementField_1/'):
           #Search for the exportPath property
           if elementMeas.get('name') == 'exportPath':
@@ -166,11 +166,11 @@ class NeutrinoBase(CodeInterfaceBase):
             exportPath = originalPath.replace("NeutrinoInput.nescene","",1)
             exportPath = exportPath + "\\Measurements\\results.csv"
             elementMeas.set('val',str(exportPath))
-                
-    
+
+
     # Now we can re-write the input file
     tree.write(currentInputFiles[index].getAbsFile())
-	
+
     return currentInputFiles
 
   def finalizeCodeOutput(self, command, output, workingDir):
@@ -186,28 +186,32 @@ class NeutrinoBase(CodeInterfaceBase):
     # create full path to the outputfile
 	# NeutrinoInput needs to be the name of the Neutrino Input file
 	# Name of results file name needs to be the same as in the createNewInput function
-    outputPath = workingDir + '\\NeutrinoInput\\Measurements\\results.csv' 
-	
+    outputPath = workingDir + '\\NeutrinoInput\\Measurements\\results.csv'
+
+	#Change the output path so RAVEN can read the output
+    newOutputPath = workingDir + "\\" + output
+
+    # check that the output file exists
+    if not os.path.exists(outputPath):
+      print('Results file does not exist. OK if during test.')
+      return newOutputPath
+
     # open original output file (the working directory is provided)
     outputFile = open(outputPath,"r+")
-	
-    #Change the output path so RAVEN can read the output
-    #output = output + ".csv"
-    newOutputPath = workingDir + "\\" + output
- 
+
     #Open the new output file so the results can be written to it and put in the form for RAVEN to read
     resultsFile = open(newOutputPath + ".csv", 'w')
 
     lines = outputFile.readlines()
-	
+
 	#Needed for RAVEN to read output
 	#These need to match RAVEN input file output names
-    resultsFile.write('time,result\n') 
+    resultsFile.write('time,result\n')
 
 	#Write Neutrino results to a new file for RAVEN
-    for line in lines: 
+    for line in lines:
       resultsFile.write(line)
-  
+
     resultsFile.close()
 
     outputFile.close()
